# DOCUMENTACAO - openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Desafio Técnico - Pulse",
        "version": "1.0",
        "contact": {
            "name": "JOAO HIGO SOUSA NUNES",
            "email": "higo.sousaa@gmail.com"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "local"
        }
    ],
    "paths": {
...
        "/cliente/endereco": {
            "get": {
                "summary": "Your GET endpoint",
                "tags": [],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "description": "",
                                    "minItems": 1,
                                    "uniqueItems": true,
                                    "items": {
                                        "type": "object",
                                        "required": [
                                            "id",
                                            "rua",
                                            "complemento",
                                            "bairro",
                                            "numero",
                                            "cep",
                                            "estado"
                                        ],
                                        "properties": {
                                            "id": {
                                                "type": "number"
                                            },
                                            "rua": {
                                                "type": "string",
                                                "minLength": 1
                                            },
                                            "complemento": {
                                                "type": "string",
                                                "minLength": 1
                                            },
                                            "bairro": {
                                                "type": "string",
                                                "minLength": 1
                                            },
                                            "numero": {
                                                "type": "number"
                                            },
                                            "cep": {
                                                "type": "number"
                                            },
                                            "estado": {
                                                "type": "string",
                                                "minLength": 1
                                            }
                                        }
                                    },
                                    "x-examples": {
                                        "example-1": [
                                            {
                                                "id": 1,
                                                "rua": "Rua Aririzal",
                                                "complemento": "Condomínio b5ap5",
                                                "bairro": "Cohama",
                                                "numero": 2,
                                                "cep": 65066260,
                                                "estado": "MA"
                                            }
                                        ]
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": [
                                            {
                                                "id": 1,
                                                "rua": "Rua Aririzal",
                                                "complemento": "Condomínio b5ap5",
                                                "bairro": "Cohama",
                                                "numero": 2,
                                                "cep": 65066260,
                                                "estado": "MA"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    }
                },
                "operationId": "get-cliente-endereco",
                "security": [
                    {
                        "Bearer token": []
                    }
                ]
            },
            "post": {
                "summary": "",
                "operationId": "post-cliente-endereco",
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "",
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "rua": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "complemento": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "bairro": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "numero": {
                                            "type": "number"
                                        },
                                        "cep": {
                                            "type": "number"
                                        },
                                        "estado": {
                                            "type": "string",
                                            "minLength": 1
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "rua",
                                        "complemento",
                                        "bairro",
                                        "numero",
                                        "cep",
                                        "estado"
                                    ],
                                    "x-examples": {
                                        "example-1": {
                                            "id": 2,
                                            "rua": "Av Daniel de la touche",
                                            "complemento": "Condomínio",
                                            "bairro": "Cohama",
                                            "numero": 2,
                                            "cep": 65066260,
                                            "estado": "MA"
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": {
                                            "id": 2,
                                            "rua": "Av Daniel de la touche",
                                            "complemento": "Condomínio",
                                            "bairro": "Cohama",
                                            "numero": 2,
                                            "cep": 65066260,
                                            "estado": "MA"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    }
                },
                "security": [
                    {
                        "Bearer token": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "",
                                "type": "object",
                                "properties": {
                                    "rua": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "numero": {
                                        "type": "number"
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "estado": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "cep": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "rua",
                                    "numero",
                                    "complemento",
                                    "bairro",
                                    "estado",
                                    "cep"
                                ],
                                "x-examples": {
                                    "example-1": {
                                        "rua": "Av Daniel de la touche",
                                        "numero": 2,
                                        "complemento": "Condomínio",
                                        "bairro": "Cohama",
                                        "estado": "MA",
                                        "cep": 65066260
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "rua": "Av Daniel de la touche",
                                        "numero": 2,
                                        "complemento": "Condomínio",
                                        "bairro": "Cohama",
                                        "estado": "MA",
                                        "cep": 65066260
                                    }
                                }
                            }
                        }
                    },
                    "description": ""
                }
            },
            "put": {
                "summary": "",
                "operationId": "put-cliente-endereco",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "",
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "number"
                                        },
                                        "rua": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "complemento": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "bairro": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "numero": {
                                            "type": "number"
                                        },
                                        "cep": {
                                            "type": "number"
                                        },
                                        "estado": {
                                            "type": "string",
                                            "minLength": 1
                                        }
                                    },
                                    "required": [
                                        "id",
                                        "rua",
                                        "complemento",
                                        "bairro",
                                        "numero",
                                        "cep",
                                        "estado"
                                    ],
                                    "x-examples": {
                                        "example-1": {
                                            "id": 2,
                                            "rua": "Av Daniel de la touche",
                                            "complemento": "Condomínio",
                                            "bairro": "Cohama",
                                            "numero": 2,
                                            "cep": 65066260,
                                            "estado": "MA"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "$ref": "#/components/responses/Nao-autorizado"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "",
                                    "type": "object",
                                    "properties": {
                                        "timeStamp": {
                                            "type": "number"
                                        },
                                        "status": {
                                            "type": "number"
                                        },
                                        "error": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "message": {
                                            "type": "string",
                                            "minLength": 1
                                        },
                                        "path": {
                                            "type": "string",
                                            "minLength": 1
                                        }
                                    },
                                    "required": [
                                        "timeStamp",
                                        "status",
                                        "error",
                                        "message",
                                        "path"
                                    ],
                                    "x-examples": {
                                        "example-1": {
                                            "timeStamp": 1621303164316,
                                            "status": 404,
                                            "error": "Não Encontrado",
                                            "message": "Endereço não encontrado no cadastro do cliente",
                                            "path": "/api/cliente/endereco"
                                        }
                                    }
                                },
                                "examples": {
                                    "example-1": {
                                        "value": {
                                            "timeStamp": 1621303164316,
                                            "status": 404,
                                            "error": "Não Encontrado",
                                            "message": "Endereço não encontrado no cadastro do cliente",
                                            "path": "/api/cliente/endereco"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer token": []
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "",
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    },
                                    "rua": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "numero": {
                                        "type": "number"
                                    },
                                    "complemento": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "bairro": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "estado": {
                                        "type": "string",
                                        "minLength": 1
                                    },
                                    "cep": {
                                        "type": "number"
                                    }
                                },
                                "required": [
                                    "id",
                                    "rua",
                                    "numero",
                                    "complemento",
                                    "bairro",
                                    "estado",
                                    "cep"
                                ],
                                "x-examples": {
                                    "example-1": {
                                        "id": 2,
                                        "rua": "Rua São Luís rei da França",
                                        "numero": 2,
                                        "complemento": "Condomínio",
                                        "bairro": "Turu",
                                        "estado": "MA",
                                        "cep": 65066260
                                    }
                                }
                            },
                            "examples": {
                                "example-1": {
                                    "value": {
                                        "id": 2,
                                        "rua": "Rua São Luís rei da França",
                                        "numero": 2,
                                        "complemento": "Condomínio",
                                        "bairro": "Turu",
                                        "estado": "MA",
                                        "cep": 65066260
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Bearer token": {
                "name": "Authorization",
                "type": "apiKey",
                "in": "header",
                "description": ""
            },
            "Perfil": {
                "type": "oauth2"
            }
        },
        "responses": {
            "Nao-autorizado": {
                "description": "Resposta padrão quando um usuário tenta acessar um recurso que ele não tem autorização.",
                "content": {
                    "application/json": {
                        "schema": {
                            "description": "",
                            "type": "object",
                            "properties": {
                                "timestamp": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "status": {
                                    "type": "number"
                                },
                                "error": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "message": {
                                    "type": "string",
                                    "minLength": 1
                                },
                                "path": {
                                    "type": "string",
                                    "minLength": 1
                                }
                            },
                            "required": [
                                "timestamp",
                                "status",
                                "error",
                                "message",
                                "path"
                            ],
                            "x-examples": {
                                "example-1": {
                                    "timestamp": "2021-05-18T01:24:42.970+00:00",
                                    "status": 403,
                                    "error": "Forbidden",
                                    "message": "Access Denied",
                                    "path": "/api/cliente/cartao"
                                }
                            }
                        },
                        "examples": {
                            "example-1": {
                                "value": {
                                    "timestamp": "2021-05-18T01:24:42.970+00:00",
                                    "status": 403,
                                    "error": "Forbidden",
                                    "message": "Access Denied",
                                    "path": "/api"
                                }
                            }
                        }
                    }
                }
            },
...
        }
    }
}